*cleanup *smooth tangents *remove intptr_t
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@762 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/client/cl_scrn.c
@@ -556,10 +556,10 @@ int entitycmpfnc (const entity_t * a, const entity_t * b) {
 	 ** all other models are sorted by model then skin
 	 */
 	if (a->model == b->model) {
-		return ((intptr_t)a->skin - (intptr_t)b->skin);
+		return ((int)a->skin - (int)b->skin);
 	}
 	else {
-		return ((intptr_t)a->model - (intptr_t)b->model);
+		return ((int)a->model - (int)b->model);
 	}
 }
